ls -a added

+++ test.sh
@@ -1,10 +1,10 @@
-if [ $(docker container ls -q --filter name=my_app_db_container) != '' ]; then
+if [ $(docker container ls -a -q --filter name=my_app_db_container) != '' ]; then
     docker container stop my_app_db_container
     docker container rm my_app_db_container
 fi
 
 
-if [ $(docker image ls -q --filter reference=my_app_db) != '' ]; then
+if [ $(docker image ls -a -q --filter reference=my_app_db) != '' ]; then
     docker image rm my_app_db
 fi
